Added better checks for integration tests, added quality-gates-backwards-compat to results

+++ test/test_onboard_service.sh
@@ -6,6 +6,16 @@ KEPTN_EXAMPLES_BRANCH=${KEPTN_EXAMPLES_BRANCH:-"master"}
 
 echo "Testing onboarding for project $PROJECT"
 
+
+# verify that the project does not exist yet via the Keptn API
+response=$(curl -X GET "${KEPTN_ENDPOINT}/controlPlane/v1/project/${PROJECT}" -H  "accept: application/json" -H  "x-token: ${KEPTN_API_TOKEN}" -k 2>/dev/null | jq -r '.projectName')
+
+if [[ "$response" == "${PROJECT}" ]]; then
+  echo "Project ${PROJECT} already exists. Please delete it using:"
+  echo "keptn delete project ${PROJECT}"
+  exit 2
+fi
+
 # test keptn create-project and onboard
 rm -rf examples
 git clone --branch ${KEPTN_EXAMPLES_BRANCH} https://github.com/keptn/examples --single-branch
@@ -13,14 +23,14 @@ cd examples/onboarding-carts
 
 echo "Creating a new project without Git upstream"
 keptn create project $PROJECT --shipyard=../../test/assets/shipyard_onboard_service.yaml
-verify_test_step $? "keptn create project ${PROJECT} - failed."
+verify_test_step $? "keptn create project ${PROJECT} failed."
 sleep 10
 
 ###########################################
 # onboard carts                           #
 ###########################################
 keptn onboard service carts --project=$PROJECT --chart=./carts
-verify_test_step $? "keptn onboard service carts - failed"
+verify_test_step $? "keptn onboard service carts failed"
 sleep 10
 
 # add functional tests
@@ -32,7 +42,7 @@ keptn add-resource --project=$PROJECT --service=carts --stage=staging --resource
 # onboard carts-db                        #
 ###########################################
 keptn onboard service carts-db --project=$PROJECT --chart=./carts-db
-verify_test_step $? "keptn onboard service carts-db - failed"
+verify_test_step $? "keptn onboard service carts-db failed"
 
 echo "Onboarding done ✓"
